refactored screen and target classes into screen.cpp and target.cpp; added powerup.cpp and powerup.hpp
