+++ lovestruck.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.4.33122.133
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "raylib-game-template", "raylib-game-template.vcxproj", "{C91AB6B6-9E36-41C7-9B67-F7B466C6A5F5}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lovestruck", "lovestruck.vcxproj", "{C91AB6B6-9E36-41C7-9B67-F7B466C6A5F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
