+++ lovestruck.vcxproj
@@ -152,31 +152,21 @@
     <ClCompile Include="src\arrow.cpp" />
     <ClCompile Include="src\button.cpp" />
     <ClCompile Include="src\cloud.cpp" />
-    <ClCompile Include="src\credits_screen.cpp" />
     <ClCompile Include="src\game.cpp" />
-    <ClCompile Include="src\ending_screen.cpp" />
-    <ClCompile Include="src\gameplay_screen.cpp" />
-    <ClCompile Include="src\logo_screen.cpp" />
     <ClCompile Include="src\main.cpp" />
-    <ClCompile Include="src\options_screen.cpp" />
     <ClCompile Include="src\player.cpp" />
+    <ClCompile Include="src\powerup.cpp" />
     <ClCompile Include="src\screen.cpp" />
     <ClCompile Include="src\target.cpp" />
-    <ClCompile Include="src\target_large.cpp" />
-    <ClCompile Include="src\target_medium.cpp" />
-    <ClCompile Include="src\target_small.cpp" />
     <ClCompile Include="src\target_spawner.cpp" />
-    <ClCompile Include="src\title_screen.cpp" />
+    <ClCompile Include="src\utils.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\animated_object.hpp" />
     <ClInclude Include="src\arrow.hpp" />
     <ClInclude Include="src\button.hpp" />
     <ClInclude Include="src\cloud.hpp" />
-    <ClInclude Include="src\credits_screen.hpp" />
-    <ClInclude Include="src\ending_screen.hpp" />
     <ClInclude Include="src\game.hpp" />
-    <ClInclude Include="src\gameplay_screen.hpp" />
     <ClInclude Include="src\include\AudioDevice.hpp" />
     <ClInclude Include="src\include\AudioStream.hpp" />
     <ClInclude Include="src\include\BoundingBox.hpp" />
@@ -215,16 +205,12 @@
     <ClInclude Include="src\include\VrStereoConfig.hpp" />
     <ClInclude Include="src\include\Wave.hpp" />
     <ClInclude Include="src\include\Window.hpp" />
-    <ClInclude Include="src\options_screen.hpp" />
     <ClInclude Include="src\player.hpp" />
+    <ClInclude Include="src\powerup.hpp" />
     <ClInclude Include="src\screen.hpp" />
-    <ClInclude Include="src\logo_screen.hpp" />
     <ClInclude Include="src\target.hpp" />
-    <ClInclude Include="src\target_large.hpp" />
-    <ClInclude Include="src\target_medium.hpp" />
-    <ClInclude Include="src\target_small.hpp" />
     <ClInclude Include="src\target_spawner.hpp" />
-    <ClInclude Include="src\title_screen.hpp" />
+    <ClInclude Include="src\utils.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
