--- src/options_screen.cpp
@@ -1,32 +1,0 @@
-#include "screen.hpp"
-#include "options_screen.hpp"
-
-//----------------------------------------------------------------------------------
-// Options Screen Functions Definition
-//----------------------------------------------------------------------------------
-
-// Options Screen Initialization logic
-void OptionsScreen::InitScreen()
-{
-    // TODO: Initialize OPTIONS screen variables here!
-    m_framesCounter = 0;
-    m_finishScreen = 0;
-}
-
-// Options Screen Update logic
-void OptionsScreen::UpdateScreen()
-{
-    // TODO: Update OPTIONS screen variables here!
-}
-
-// Options Screen Draw logic
-void OptionsScreen::DrawScreen()
-{
-    // TODO: Draw OPTIONS screen here!
-}
-
-// Options Screen Unload logic
-void OptionsScreen::UnloadScreen()
-{
-    // TODO: Unload OPTIONS screen variables here!
-}
