--- src/target_medium.cpp
@@ -1,24 +1,0 @@
-#include "target_medium.hpp"
-
-TargetMedium::TargetMedium(
-	raylib::Vector2 pos,
-	raylib::Texture2DUnmanaged& aliveTexture,
-	raylib::Texture2DUnmanaged& endingTexture
-)	: Target(pos, aliveTexture, endingTexture)
-{
-	m_maxFrames = 3;
-	m_width = m_texture.width / m_maxFrames;
-	m_height = m_texture.height;
-	m_velocity = 250.0f;
-	m_scale = 0.6f;
-	m_health = 2; // 2 hits to destroy
-	m_score = 2;
-}
-
-TargetMedium::~TargetMedium() {};
-
-void TargetMedium::onHit()
-{
-	Target::onHit();
-	this->m_frame++;
-}
