--- src/target_medium.hpp
@@ -1,15 +1,0 @@
-#pragma once
-
-#include "target.hpp"
-
-class TargetMedium : public Target
-{
-public:
-	TargetMedium(
-		raylib::Vector2 pos, 
-		raylib::Texture2DUnmanaged& aliveTexture,
-		raylib::Texture2DUnmanaged& endingTexture
-	);
-	~TargetMedium();
-	void onHit();
-};
